+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## [unreleased]
+* made display command prettier
 * print time since last stop
 
 ## [0.7.0] - 2022-04-24